% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_processingQC.R
\name{addUpdatedCDFenv}
\alias{addUpdatedCDFenv}
\title{Load an updated cdf environment from BrainArray}
\usage{
addUpdatedCDFenv(Data, species = NULL, type = "ENSG")
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{species}{(Status: required, Default: NULL) The species associated with the chip type (datatype: character).
Possible values:
- Abbreviated:    "Ag", "At", "Bt", "Ce", "Cf", "Dr", "Dm", "Gg", "Hs", "MAmu", "Mm", "Os", "Rn", "Sc", "Sp", "Ss"
- Full names:    "Anopheles gambiae", "Arabidopsis thaliana", "Bos taurus", "Caenorhabditis elegans",
 "Canis familiaris", "Danio rerio", "Drosophila melanogaster", "Gallus gallus", "Homo sapiens",
 "Macaca mulatta", "Mus musculus", "Oryza sativa", "Rattus norvegicus", "Saccharomyces cerevisiae",
 "Schizosaccharomyces pombe", "Sus scrofa".}

\item{type}{(Status: optional, Default: "ENSG") The type of custom cdf requested
This parameter indicates the database based on which
an updated cdf environment should be selected (datatype: character).
Possible values are:
"ENTREZG", "REFSEQ", "ENSG", "ENSE", "ENST", "VEGAG", "VEGAE", "VEGAT", "TAIRG", "TAIRT",
"UG", "MIRBASEF", "MIRBASEG".}
}
\value{
The object (of class AffyBatch) with a updated cdf annotation assigned if found
}
\description{
This function (from functions_processing) tries to find and
load an updated cdf environment from \href{http://brainarray.mbni.med.umich.edu/brainarray}{BrainArray} for the current chiptype.
If this is not successful, a warning will be generated and the raw
data object returned as is, i.e. with the cdf annotation that it already had,
if any. Note that the IDs given to the reporters are artificially
created by adding “_at” as a postfix to the ID from the entry
in the database that the probeset corresponds to (c.f. also createNormDataTable).
Note also that reannotation takes places at a probe level, not at a probeset level.
This means that completely new probesets are constructed, not necessarily equal in size.
}
\examples{
#By default, the script will call, if customCDF is TRUE
#(from within the \link[ArrayAnalysis]{normalizeData}
#and \link[ArrayAnalysis]{computePMAtable} functions,
#leaving the original Data intact and using Data.copy to further process):
#Data.copy <- Data
#Data.copy <- addUpdatedCDFenv(Data.copy, species, CDFtype)
}
