% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{array.image}
\alias{array.image}
\title{Create several types of spatial images}
\usage{
array.image(
  Data,
  pcut = NULL,
  relative = TRUE,
  symm = relative,
  balance = relative,
  quantitative = relative,
  col.mod = 1,
  postfix = "",
  arrays = NULL,
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch or ExpressionSet)}

\item{pcut}{(Status: optional, Default:NULL. This means that for relative plots a
value of 0.001 will be used, and for other plots values of c(0.01, 0.05).)
Either a numeric value or a vector of two numeric
values, both in the interval [0, 0.5], used a
color saturation limits as a percentage of the
data range. If one value is provided, this is
taken both a lower and upper saturation limit. (datatype: numeric)}

\item{relative}{(Status: optional, Default:TRUE) Is the plot to be created
a plot of each array relative to the median array, or not (i.e. an
absolute plot)? (datatype: logical)}

\item{symm}{(Status: optional, Default: TRUE if "relative" is TRUE,
FALSE otherwise.) Should a symmetric color scale be used? (datatype: logical)}

\item{balance}{(Status: optional, Default:Default: TRUE if "relative" is TRUE,
FALSE otherwise.) Should the arrays first be balanced for their
average intensities? (datatype: logical)}

\item{quantitative}{(Status: optional, Default:TRUE if "relative"is TRUE,
FALSE otherwise. Setting "quantitative" to TRUE has no effect if "relative" is FALSE,
a warning will be produced.) Should the plot be quantitative or qualitative
(i.e. only indicate the sign of the value)? (datatype: logical)}

\item{col.mod}{(Status: optional, Default:1) A numeric value used as a
modifier for the color range. A value of 1 means no modification
(linear), smaller leads to faster saturation, larger to slower saturation. (datatype: numeric)}

\item{postfix}{(Status: optional, Default:"") String to be attached to the file names produced. (datatype: character)}

\item{arrays}{(Status: optional, Default:NULL (in which case
all arrays are plotted)) Which arrays are to be plotted. NOTE: for
relative plot types, the median array is still computed using all arrays
in the dataset. (datatype: numeric)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}
}
\value{
Images containing each of the requested plots, one file per six arrays.
Naming is "virtual_array_plots" followed by the postfix given by the user.
}
\description{
This function (from functions_imagesQC.R) creates several types of
spatial images. It does not make use of a PLMset object and is called
when the PLM images cannot be computed due to time or memory constaints.
Also, it offers more flexibility. By default, it creates
a relative intensity plot versus the median array on a blue to
red colour scale. Note that the median array will always be computed
based on the complete data set, also when plots for a subset of
arrays are requested. If the median array has to be computed on
subsets (e.g. on experimental group), a data object with only
those arrays can be provided to the function (without setting
the arrays parameter). The color ranges will saturate at pcut
percentage(s) of the data range. The color ranges can be modified
by tuning col.mod. By default, for the relative plots, the arrays
are first balanced for their overall intensity and a symmetrical
color range is used. When there is less than 6 arrays, the median
array is not used. Intensities are plotted using a virtual symetric
color scale, from blue to red.
}
\examples{
# There are two default calls by the script:
# 1/ relative intensity plot versus the median array on a blue to red colour scale:
# array.image(rawData)
# 2/ absolute intensity plot when there are less than 6 arrays in the dataset:
# array.image(rawData,relative=FALSE,col.mod=4,symm=TRUE)
# Other calls could be made, such as:
# spatial intensity plot on a virtual colour scale (red to yellow)
# array.image(rawData,relative=FALSE)
# signs of the relative intensities versus the median array
# (e.g. lower or higher) with blue as lower and red as higher.
# array.image(rawData,quantitative=FALSE)
# similar to the default plot, but not balanced within arrays
# array.image(rawData,balance=FALSE)
# or: per experimental group
# for (i in levels(experimentFactor)) {
# array.image(rawData[, experimentFactor == i], postfix = paste("_",i), balance=FALSE)
# }
}
