% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{boxplotFun}
\alias{boxplotFun}
\title{Create an image with intensity boxplots for all arrays}
\usage{
boxplotFun(
  Data,
  experimentFactor = NULL,
  plotColors = NULL,
  legendColors = NULL,
  normMeth = "",
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24,
  MAXARRAY = 41
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{experimentFactor}{(Status: required, Default:NULL) The factor of groups. (datatype: factor)}

\item{plotColors}{(Status: required, Default:NULL) Vector of colors assigned to each array. (datatype: character)}

\item{legendColors}{(Status: required, Default:NULL) Vector of colors assigned to each experimental group. (datatype: character)}

\item{normMeth}{(Status: required when Data is a normalized data object, Default:"")
String indicating the normalization method used (see
\link[ArrayAnalysis]{normalizeData} function for more information on the possible values). (datatype: character)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}

\item{MAXARRAY}{(Status: optional, Default:41) threshold to adapt the image to the number of arrays (datatype: number)}
}
\value{
A PNG image of the raw or normalized boxplots of the arrays, called ‘DataBoxplot’
}
\description{
This function (from functions_imagesQC.R) creates an image
with intensity boxplots for all arrays in the raw or normalized
dataset (depending on the object passed). It calls the boxplot
function applied to an AffyBatch object.
}
\examples{
# By default, before the normalization the script will call:
# boxplotFun(Data=rawData, experimentFactor, plotColors, legendColors)
# and after normalization:
# boxplotFun(Data=normData, experimentFactor, plotColors, legendColors, normMeth=normMeth)
}
