% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{clusterFun}
\alias{clusterFun}
\title{Create a hierarchical clustering dendrogam}
\usage{
clusterFun(
  Data,
  experimentFactor = NULL,
  clusterOption1 = "pearson",
  clusterOption2 = "ward.D2",
  normMeth = "",
  plotColors = NULL,
  legendColors = NULL,
  plotSymbols = NULL,
  legendSymbols = NULL,
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24,
  MAXARRAY = 41
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{experimentFactor}{(Status: required, Default:NULL) The factor of groups. (datatype: factor)}

\item{clusterOption1}{(Status: optional, Default:"Pearson") String indicating
the distance function to be used. Possible values are "pearson", “spearman”,
or “euclidean”. (datatype: character)}

\item{clusterOption2}{(Status: optional, Default:"ward") String indicating
the hierarchical clustering function to be used. Possible values are "ward",
"single", "complete", "average", "mcquitty", "median" or "centroid". (datatype: character)}

\item{normMeth}{(Status: required when Data is a normalized data object, Default:"")
String indicating the normalization method used (see
\link[ArrayAnalysis]{normalizeData} function for more information on the possible values). (datatype: character)}

\item{plotColors}{(Status: required, Default:NULL) Vector of colors assigned to each array. (datatype: character)}

\item{legendColors}{(Status: required, Default:NULL) Vector of colors assigned to each experimental group. (datatype: character)}

\item{plotSymbols}{(Status: required, Default:NULL) symbol assigned to each array. (datatype: number)}

\item{legendSymbols}{(Status: required, Default:NULL) symbol assigned to each experimental group. (datatype: number)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}

\item{MAXARRAY}{(Status: optional, Default:41) threshold to adapt the image to the number of arrays (datatype: number)}
}
\value{
A PNG image with the clustering dendrogram of the arrays.
Naming is ‘DataCluster’ followed by the name of the distance function used
}
\description{
This function (from functions_imagesQC.R) creates a hierarchical
clustering dendrogam of the arrays in the raw or normalized dataset
(depending on the object passed). When the dataset consists of less
than three arrays, no dendrogram is generated and a warning is given.
}
\examples{
# By default, before the normalization the script will call:
# clusterFun(Data=rawData, clusterOption1=clusterOption1,
# clusterOption2=clusterOption2)
# and after normalization:
# clusterFun(Data=normData, clusterOption1=clusterOption1,
# clusterOption2=clusterOption2, normMeth=normMeth)
}
