% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_processingQC.R
\name{computePMAtable}
\alias{computePMAtable}
\title{Prepare and return a table of PMA calls}
\usage{
computePMAtable(Data, customCDF = TRUE, species = NULL, CDFtype = NULL)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{customCDF}{(Status: optional, Default: TRUE) Should annotation of the
chip be updated before normalizing the data (and building the probesets
out of the separate probes)? If requested, this is done using
\href{http://brainarray.mbni.med.umich.edu/brainarray}{BrainArray} updated cdf environments,
c.f. \link[ArrayAnalysis]{addUpdatedCDFenv} (datatype: logical)}

\item{species}{(Status: required when customCDF is TRUE, c.f.
\link[ArrayAnalysis]{addUpdatedCDFenv}, Default: NULL)The species associated with the chip type. (datatype: character)}

\item{CDFtype}{(Status: required when customCDF is TRUE, c.f.
\link[ArrayAnalysis]{addUpdatedCDFenv}, Default: NULL) The type of custom cdf requested. (datatype: character)}
}
\value{
A data.frame table called ‘PMAtable’
containing the PMA values for each probeset and array
}
\description{
This function (from functions_processing) computes a table of Absent (A),
Marginal (M), Present (P) calls based on the MAS5 function (affy package).
This function will only work for chiptypes that have mismatch probes and
for which the detection.p.val (
\href{http://bioinformatics.picr.man.ac.uk/research/software/simpleaffy/index.html}{simpleaffy}
Bioconductor package) function
that it calls works. A try construction will be used and if no table can
be created a warning is given. Note that this function will always use
the MAS5 algorithm, regardless of the normalization method used in
the \link[ArrayAnalysis]{normalizeData} function. In case customCDF is TRUE, annotation is
updated using \href{http://brainarray.mbni.med.umich.edu/brainarray}{BrainArray} custom cdf environments, before proceeding with
the normalization (and summarisation of probe expressions into probeset
expressions). To update the annotation, a sub call is made to the
\link[ArrayAnalysis]{addUpdatedCDFenv} function.
}
\examples{
#By default, the script will call, if customCDF is TRUE:
#PMAtable <- computePMAtable(rawData,customCDF,species,CDFtype)
#or, if customCDF is FALSE:
#PMAtable <- computePMAtable(rawData,customCDF)
#After this call, the main script will save the result
#to a tab-delimited text file, called PMAtable.txt
}
