% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{computeStatistics}
\alias{computeStatistics}
\title{Compute the statistics}
\usage{
computeStatistics(
  normDataTable,
  descriptionFile,
  defaultContr = TRUE,
  matfileName = NULL,
  keepAnnotation = FALSE
)
}
\arguments{
\item{normDataTable}{normalized data object (Status=required)}

\item{descriptionFile}{(Status: required) The data.frame containing the description
file information (column 1: file names; column 2: names
to be used in the plots; column 3: experimental
groups the samples belong to)(datatype: data.frame)}

\item{defaultContr}{(datatype=logical, Default=TRUE)}

\item{matfileName}{(datatype=character, Default=NULL)}

\item{keepAnnotation}{(datatype=logical, Default=FALSE)}
}
\value{
the contrast matrix file
}
\description{
Compute the statistics
}
\examples{
#example here
}
