% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{controlPlots}
\alias{controlPlots}
\title{Create an image of the scale factors of the array}
\usage{
controlPlots(
  Data,
  plotColors = NULL,
  experimentFactor = NULL,
  legendColors = NULL,
  affxplots = TRUE,
  boxplots = TRUE,
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24,
  MAXARRAY = 41
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{plotColors}{(Status: required, Default:NULL) Vector of colors assigned to each array. (datatype: character)}

\item{experimentFactor}{(Status: required, Default:NULL) The factor of groups. (datatype: factor)}

\item{legendColors}{(Status: required, Default:NULL) Vector of colors assigned to each experimental group. (datatype: character)}

\item{affxplots}{(Status: optional, Default:TRUE) Does the AFFX expression profiles have to be plotted? (datatype: logical)}

\item{boxplots}{(Status: optional, Default:TRUE) Does the AFFX and other controls boxplots have to be plotted? (datatype: logical)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}

\item{MAXARRAY}{(Status: optional, Default:41) threshold to adapt the image to the number of arrays (datatype: number)}
}
\value{
Two images. A PNG image of the expression profiles of the affx controls, called "Profiles_affx_controls"
A PNG image of the boxplots of the affx controls, called "Boxplots_affx_controls"
}
\description{
This function (from functions_imagesQC.R) creates images of
the expression values of the affx controls, if present. One
image shows the expression profiles over all samples, for each
control separately. The other image shows the boxplots of all
controls (together), for each sample.
}
\examples{
# controlPlots(rawData,plotColors)
}
