% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{correlFun}
\alias{correlFun}
\title{Create an image of the intensity correlation values}
\usage{
correlFun(
  Data,
  clusterOption1 = "pearson",
  clusterOption2 = "ward.D2",
  normMeth = "",
  experimentFactor = NULL,
  legendColors = NULL,
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24,
  MAXARRAY = 41
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{clusterOption1}{(Status: optional, Default:"Pearson") String indicating
the distance function to be used. Possible values are "pearson", “spearman”,
or “euclidean”. (datatype: character)}

\item{clusterOption2}{(Status: optional, Default:"ward") String indicating
the hierarchical clustering function to be used. Possible values are "ward",
"single", "complete", "average", "mcquitty", "median" or "centroid". (datatype: character)}

\item{normMeth}{(Status: required when Data is a normalized data object, Default:"")
String indicating the normalization method used (see
\link[ArrayAnalysis]{normalizeData} function for more information on the possible values). (datatype: character)}

\item{experimentFactor}{(Status: required, Default:NULL) The factor of groups. (datatype: factor)}

\item{legendColors}{(Status: required, Default:NULL) Vector of colors assigned to each experimental group. (datatype: character)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}

\item{MAXARRAY}{(Status: optional, Default:41) threshold to adapt the image to the number of arrays (datatype: number)}
}
\value{
A PNG image of the intensity correlation values of the arrays, called "DataArrayCorrelationPlot"
}
\description{
This function (from functions_imagesQC.R) creates an image of the
intensity correlation values of the arrays in the raw or normalized dataset
(depending on the object passed). It calls correlationPlot
(affyQCReport Bioconductor package).
}
\examples{
# By default, before the normalization the script will call:
# correlFun(Data=rawData)
# and after normalization:
# correlFun(Data=normData, normMeth=normMeth)
}
