% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{coverAndKeyPlot}
\alias{coverAndKeyPlot}
\title{Plot a cover sheet}
\usage{
coverAndKeyPlot(description = NULL, refName = "", WIDTH = 1000, HEIGHT = 1414)
}
\arguments{
\item{description}{(Status: required) The data.frame containing the description
file information (column 1: file names; column 2: names
to be used in the plots; column 3: experimental
groups the samples belong to)(datatype: data.frame)}

\item{refName}{(Status: optional, Default:"") dataset name. It is deduced
from the name of the zip file containing the CEL files when used from
arrayanalysis.org or as a GenePattern module.(datatype: character)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}
}
\value{
Two files, a file called ‘Cover_1’ that can be used as an opening image.And,
file(s) called ‘Description’, if needed followed by an index number,that
represent(s) the information as given in the input parameter.
}
\description{
This function (from functions_imagesQC.R) plots a cover sheet,
and one or more key sheet indicating the links between CEL file names,
array names used in the plots, and to which experimental group they
belong based on the description file provided by the user, which has
been loaded into the description variable earlier in the main script,
passed as an argument (arrayGroup). For the key sheets, one sheet will
be created for every 35 arrays in the experiment.
}
\examples{
# By default, the script will call:
# coverAndKeyPlot(description)
}
