% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{createCutOffTab}
\alias{createCutOffTab}
\title{For each comparison, create a cutoff table (P-value, LogFC and Average expression)}
\usage{
createCutOffTab(
  files,
  cutOffPval = NULL,
  cutOfflogFC = NULL,
  cutOffAveExpr = NULL
)
}
\arguments{
\item{files}{(Status=required) vector of fit-model file names}

\item{cutOffPval}{(default=NULL)}

\item{cutOfflogFC}{(default=NULL)}

\item{cutOffAveExpr}{(default=NULL)}
}
\value{
several files (for each comparison)
}
\description{
For each comparison, create a cutoff table (P-value, LogFC and Average expression)
}
\examples{
#example here
}
