% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{createFCHist}
\alias{createFCHist}
\title{For each comparison table, create a Fold Change histogram plot}
\usage{
createFCHist(files)
}
\arguments{
\item{files}{(Status=required) vector of fit-model file names}
}
\value{
a PNG file for the histogram plot
}
\description{
For each comparison table, create a Fold Change histogram plot
}
\examples{
#example here
}
