% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_processingQC.R
\name{createNormDataTable}
\alias{createNormDataTable}
\title{Prepare and export the normalized data table}
\usage{
createNormDataTable(normData, customCDF = NULL, species = NULL, CDFtype = NULL)
}
\arguments{
\item{normData}{(Status: required) The normalized data object (datatype: ExpressionSet)}

\item{customCDF}{(Status: optional, Default: TRUE) Should annotation of the
chip be updated before normalizing the data (and building the probesets
out of the separate probes)? If requested, this is done using
\href{http://brainarray.mbni.med.umich.edu/brainarray}{BrainArray} updated cdf environments,
c.f. \link[ArrayAnalysis]{addUpdatedCDFenv} (datatype: logical)}

\item{species}{(Status: required when customCDF is TRUE, c.f.
\link[ArrayAnalysis]{addUpdatedCDFenv}, Default: NULL)The species associated with the chip type. (datatype: character)}

\item{CDFtype}{(Status: required when customCDF is TRUE, c.f.
\link[ArrayAnalysis]{addUpdatedCDFenv}, Default: NULL) The type of custom cdf requested. (datatype: character)}
}
\value{
A data.frame table with normalized data, and possible extra annotation.
}
\description{
This function (from functions_processing) prepares a table suitable
for visualisation and saving to disk from a normalized data object.
In case of use of an updated cdf annotation (customCDF is TRUE), it
will remove the artificial “_at” postfixes from all probeset IDs,
apart from the affx controls, in order to get the real IDs from the
database used to create the update. Also, in case an updated cdf
environment based on Ensemble Gene ID (“ENSG”) has been used,
the function tries to connect to \href{http://www.biomart.org/}{BioMart} (using the
\href{http://www.bioconductor.org/packages/release/bioc/html/biomaRt.html}{biomaRt} BioConductor package) in order to add two extra
columns of information to the table: the common gene name,
and a gene description. This will (for now) not be done for
other updated cdf types, as there is no one-to-one mapping
between BioMart entries and these IDs or it has not been
sufficiently tested. Note that the BioMart connection may
sometimes not be established (e.g. if the service is down or busy),
it may in such cases be worthwhile to try again.
}
\examples{
#By default, the script will call:
#normDataTable <- createNormDataTable(normData, customCDF, species, CDFtype)
#After this function has been called, the normDataTable object is
#saved to a tab-delimited text file by the main script.
}
