% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{createPvalHist}
\alias{createPvalHist}
\title{For each comparison table, create a P-value histogram plot}
\usage{
createPvalHist(files)
}
\arguments{
\item{files}{(Status=required) vector of fit-model file names}
}
\value{
a PNG file for the histogram plot
}
\description{
For each comparison table, create a P-value histogram plot
}
\examples{
#example here
}
