% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{createPvalTab}
\alias{createPvalTab}
\title{Create significant pvalue table}
\usage{
createPvalTab(
  files,
  pvaluelist = c(0.001, 0.01, 0.05, 0.1),
  adjpvaluelist = 0.05,
  foldchangelist = c(1.1, 1.5, 3),
  html = FALSE
)
}
\arguments{
\item{files}{(Status=required) vector of fit-model file names}

\item{pvaluelist}{(datatype=vector, Default=c(0.001,0.01,0.05, 0.1))}

\item{adjpvaluelist}{(datatype=numeric, Default=0.05)}

\item{foldchangelist}{(datatype=vector, Default=c(1.1,1.5,3))}

\item{html}{(datatype=logical, Default=FALSE)}
}
\value{
a summary table file
}
\description{
Create significant pvalue table
}
\examples{
#example here
}
