% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{createVennPlot}
\alias{createVennPlot}
\title{create a venn diagram plot}
\usage{
createVennPlot(fit)
}
\arguments{
\item{fit}{(Status=required) fit object}
}
\value{
a PNG file for the plot
}
\description{
create a venn diagram plot
}
\examples{
#example here
}
