% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{createVolcanoPlot}
\alias{createVolcanoPlot}
\title{create a volcano plot}
\usage{
createVolcanoPlot(fit)
}
\arguments{
\item{fit}{(Status=required) fit object}
}
\value{
a PNG file for the plot
}
\description{
create a volcano plot
}
\examples{
#example here
}
