% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{defaultMatrix}
\alias{defaultMatrix}
\title{Default comparison (for less than 5 experimental factors)}
\usage{
defaultMatrix(design)
}
\arguments{
\item{design}{(Status=required)}
}
\value{
a dataframe for the contrast matrix
}
\description{
Default comparison (for less than 5 experimental factors)
}
\examples{
#example here
}
