% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{densityFun}
\alias{densityFun}
\title{Create an image with a density curve of the intensities}
\usage{
densityFun(
  Data,
  plotColors = NULL,
  normMeth = "",
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24,
  MAXARRAY = 41
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{plotColors}{(Status: required, Default:NULL) Vector of colors assigned to each array. (datatype: character)}

\item{normMeth}{(Status: required when Data is a normalized data object, Default:"")
String indicating the normalization method used (see
\link[ArrayAnalysis]{normalizeData} function for more information on the possible values). (datatype: character)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}

\item{MAXARRAY}{(Status: optional, Default:41) threshold to adapt the image to the number of arrays (datatype: number)}
}
\value{
A PNG image of the raw or normalized density plots of the arrays, called "DensityHistogram"
}
\description{
This function (from functions_imagesQC.R) creates an image with
a density curve of the intensities for all arrays in the raw or
normalized dataset (depending on the object passed).
}
\examples{
# By default, before the normalization the script will call:
# densityFun(Data=rawData, plotColors)
# and after normalization:
# densityFun(Data=normData, plotColors, normMeth=normMeth)
}
