% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{enterMatrix}
\alias{enterMatrix}
\title{Advanced comparison: enter custom contrasts}
\usage{
enterMatrix(matfileName, groupNames)
}
\arguments{
\item{matfileName}{(Status=required)}

\item{groupNames}{(Status=required)}
}
\value{
a dataframe for the contrast matrix
}
\description{
Advanced comparison: enter custom contrasts
}
\examples{
#example here
}
