% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_processingQC.R
\name{getArrayType}
\alias{getArrayType}
\title{Determine array type used}
\usage{
getArrayType(Data)
}
\arguments{
\item{Data}{An AffyBatch object}
}
\value{
A string indicating the type of the current chip,
either “PMMM” for chips with perfect match and mismatch probes,
or “PMonly” for chips with perfect match probes only
}
\description{
Determine array type used
}
\examples{
#aType <- getArrayType(rawData)
}
