% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{maFun}
\alias{maFun}
\title{Create MA plots for each array versus the median array for the raw or normalized dataset}
\usage{
maFun(
  Data,
  experimentFactor = NULL,
  perGroup = FALSE,
  normMeth = "",
  aType = NULL,
  WIDTH = 1000,
  HEIGHT = 1414,
  MAXARRAY = 41
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch or ExpressionSet)}

\item{experimentFactor}{(Status: required when perGroup is TRUE, Default:NULL) The factor of groups. (datatype: factor)}

\item{perGroup}{(Status: optional, Default:FALSE) Are MA plots to be made for each experimental group separately or not? (datatype: logical)}

\item{normMeth}{(Status: required when Data is a normalized data object, Default:"")
String indicating the normalization method used (see
\link[ArrayAnalysis]{normalizeData} function for more information on the possible values). (datatype: character)}

\item{aType}{(Status: optional, Default:NULL) String indicating the type of
the current chip, either “PMMM” for chips with perfect match and
mismatch probes, or “PMonly” for chips with perfect match probes only.
Required when Data is a raw data object. (datatype: character)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{MAXARRAY}{(Status: optional, Default:41) threshold to adapt the image to the number of arrays (datatype: number)}
}
\value{
Images of the MA plots of each array versus the median array, each file
contains MA plots for six arrays. The files contain the string ‘MAplot” and a
number if more than one are needed; in case of groupwise computation, the name
of the group is also included in the filename.
}
\description{
This function (from functions_imagesQC.R) creates MA plots for each
array versus the median array for the raw or normalized dataset.
The median array is computed for the whole data set (if perGroup is FALSE)
of per experimental group (perGroup is TRUE). In the script this setting
will depend on the setting op the MAOption1 parameter, which can have
the values “dataset” or “group”.
}
\examples{
# By default, before the normalization the script will call:
# maFun(Data=rawData, experimentFactor, perGroup=(MAOption1=="group"), aType=aType)
# and after normalization:
# maFun(Data=normData, experimentFactor, perGroup=(MAOption1=="group"), normMeth=normMeth)
}
