% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_processingQC.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize the data set}
\usage{
normalizeData(
  Data,
  normMeth = "",
  perGroup = FALSE,
  experimentFactor = NULL,
  customCDF = TRUE,
  species = NULL,
  CDFtype = NULL,
  aType = NULL,
  isOligo = FALSE,
  WIDTH = 1000,
  HEIGHT = 1414
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{normMeth}{(Status: required, Default: "") String indicating the
normalization method used. Possible values: RMA, GCRMA, PLIER or none. (datatype: character)}

\item{perGroup}{(Status: optional, Default: FALSE) Should normalization
be performed per experimental group (e.g. when global differences
are expected between groups) or for the dataset as a whole? (datatype: logical)}

\item{experimentFactor}{(Status: required if perGroup is TRUE , Default: NULL)
The factor of groups. (datatype: factor)}

\item{customCDF}{(Status: optional, Default: TRUE) Should annotation of the
chip be updated before normalizing the data (and building the probesets
out of the separate probes)? If requested, this is done using
\href{http://brainarray.mbni.med.umich.edu/brainarray}{BrainArray} updated cdf environments,
c.f. \link[ArrayAnalysis]{addUpdatedCDFenv} (datatype: logical)}

\item{species}{(Status: required when customCDF is TRUE, c.f.
\link[ArrayAnalysis]{addUpdatedCDFenv}, Default: NULL) The species
associated with the chip type. (datatype: character)}

\item{CDFtype}{(Status: required when customCDF is TRUE, c.f.
\link[ArrayAnalysis]{addUpdatedCDFenv}, Default: NULL) The type
of custom cdf requested. (datatype: character)}

\item{aType}{(Status: optional, Default: NULL) String indicating
the type of the current chip, either “PMMM” for chips with perfect
match and mismatch probes, or “PMonly” for chips with
perfect match probes only. Required if normMeth is “PLIER”. (datatype: character)}

\item{isOligo}{(Status: optional, Default:FALSE) is Oligo array or not (datatype: logical)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}
}
\value{
A normalized data object (datatype: ExpressionSet).
It also returns A reference sheet (PNG file) indicating the cdf
annotation (cdfName) used, the normalization method used,
and if this is the case stating that normalization has been
performed per experimental group
}
\description{
This function (from functions_processing) normalizes the data
in the AffyBatch object provided. Currently, GCRMA, RMA, and
PLIER normalization are supported. For GCRMA, fast normalization
is not used, as this gives unreliable results. For PLIER,
justPlier (\href{http://www.bioconductor.org/help/bioc-views/release/bioc/html/plier.html}{plier} Bioconductor package) is used, with the
"together" option for arrays having perfect match and mismatch probes,
and the "PMonly" option for arrays with perfect match probes only.
When normalization per experimental group is selected, the function
makes sure that still one normalized data object including all arrays
is returned. In case customCDF is TRUE, annotation is updated
using \href{http://brainarray.mbni.med.umich.edu/brainarray}{BrainArray}
custom cdf environments, before proceeding with
the normalization (and summarisation of probe expressions into
probeset expressions). To update the annotation, a sub call is
made to the \link[ArrayAnalysis]{addUpdatedCDFenv} function.
}
\examples{
#By default, the script will call, if customCDF is TRUE:
#normData <- normalizeData(rawData, normMeth,
#perGroup=(normOption1=="group"), experimentFactor, customCDF,
#species, CDFtype)
#or, if customCDF is FALSE:
#normData <- normalizeData(rawData, normMeth,
#perGroup=(normOption1=="group"), experimentFactor, customCDF)
}
