% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{percPresPlot}
\alias{percPresPlot}
\title{Creates an image of the percentage present values of each array}
\usage{
percPresPlot(
  Data,
  quality = NULL,
  experimentFactor = NULL,
  plotColors = NULL,
  legendColors = NULL,
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24,
  MAXARRAY = 41
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{quality}{(Status: optional, Default:NULL) object obtained by
calling the qc function (simpleaffy).When not provided, it
is computed within the function.(datatype: QCStats)}

\item{experimentFactor}{(Status: required, Default:NULL) The factor of groups. (datatype: factor)}

\item{plotColors}{(Status: required, Default:NULL) Vector of colors assigned to each array. (datatype: character)}

\item{legendColors}{(Status: required, Default:NULL) Vector of colors assigned to each experimental group. (datatype: character)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}

\item{MAXARRAY}{(Status: optional, Default:41) threshold to adapt the image to the number of arrays (datatype: number)}
}
\value{
A PNG image of percent present values, called "RawDataPercentPresentPlot"
}
\description{
This function (from functions_imagesQC.R) creates an image of the
percentage present values of each array based on the qc
(\href{http://bioinformatics.picr.man.ac.uk/research/software/simpleaffy/index.html}{simpleaffy} Bioconductor package) function, which generally
only work for chiptypes with perfect match and mismatch probes,
but even not for all of those. As such, when these statistics
are not provided as parameters, trys are used in this function
to compute them internally. Values for which the try fails are
not computed, but the script will continue after giving a warning.
}
\examples{
# percPresPlot(rawData, quality=quality, experimentFactor, plotColors, legendColors)
}
