% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{plotArrayLayout}
\alias{plotArrayLayout}
\title{Create an image of the layout of the current chiptype}
\usage{
plotArrayLayout(
  Data,
  aType = NULL,
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch or ExpressionSet)}

\item{aType}{(Status: optional, Default:NULL) String indicating the type of
the current chip, either “PMMM” for chips with perfect match and
mismatch probes, or “PMonly” for chips with perfect match probes only.
Required when Data is a raw data object. (datatype: character)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}
}
\value{
A PNG image of the expression profiles of the affx controls, called "Array_layout_plot"
}
\description{
This function (from functions_imagesQC.R) creates an image of the layout
of the current chiptype. Thus, this plot does not plot any data, but shows
where control and regular perfect match (PM) and (if applicable)
mismatch (MM) probes are present on the array. Note: due to resolution issues,
banding may seem different from the real situation,
e.g. normally on classical chiptypes, PM and MM are present
in alternate lines, but patterns may appear due to image resolution.
This function tries to load annotation libraries, depending on the chiptype.
}
\examples{
# plotArrayLayout(rawData,aType)
}
