% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{PNdistrPlot}
\alias{PNdistrPlot}
\title{Create an image with boxplots of positive and negative control intensities}
\usage{
PNdistrPlot(Data, WIDTH = 1000, HEIGHT = 1414, POINTSIZE = 24)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}
}
\value{
A PNG image of boxplots of positive and negative control intensities, called "RawDataPosNegDistribPlot"
}
\description{
This function (from functions_imagesQC.R) creates an image with
boxplots of positive and negative control intensities for each array.
It calls the borderQC1 function (\href{http://www.bioconductor.org/help/bioc-views/release/bioc/html/affyQCReport.html}{affyQCReport} Bioconductor package).
}
\examples{
# PNdistrPlot(rawData)
}
