% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{PNposPlot}
\alias{PNposPlot}
\title{Create an image showing the centres of intensity of the positive and negative border elements}
\usage{
PNposPlot(Data, WIDTH = 1000, HEIGHT = 1414, POINTSIZE = 24)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}
}
\value{
An image of centres of intensity of positive and negative border
elements for each array, called "RawDataPosNegPositionPlot"
}
\description{
This function (from functions_imagesQC.R) creates an image showing the centres of intensity of the positive and negative border elements for each array. It calls the borderQC2 function ( affyQCReport Bioconductor package).
}
\examples{
# PNposPlot(rawData)
}
