% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{QCtablePlot}
\alias{QCtablePlot}
\title{Compute and plot a table of QC statistics}
\usage{
QCtablePlot(
  Data,
  quality = NULL,
  sprep = NULL,
  lys = NULL,
  samplePrep = TRUE,
  ratio = TRUE,
  hybrid = TRUE,
  percPres = TRUE,
  bgPlot = TRUE,
  scaleFact = TRUE,
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{quality}{(Status: optional, Default:NULL) object obtained by
calling the qc function (simpleaffy).When not provided, it
is computed within the function.(datatype: QCStats)}

\item{sprep}{(Status: optional, Default:NULL) A matrix of 3’probe
intensities for dap, thr, lys, and phe, taken from an object
of class YAQCStats (yaqc function, yaqcaffy). When not provided,
it is computed within the function. (datatype: matrix)}

\item{lys}{(Status: optional, Default:NULL) Matrix of A, M, P calls
for the 3’ probeset of Lys on each array, based on results from
the detection.p.val function (simpleaffy) . When not provided,
it is computed within the function. (datatype: matrix)}

\item{samplePrep}{(Status: optional, Default:TRUE) Does the table
have to contain sample preparation QC statistics? (datatype: logical)}

\item{ratio}{(Status: optional, Default:TRUE) Does the table have to contain 3’/5’ ratio statistics? (datatype: logical)}

\item{hybrid}{(Status: optional, Default:TRUE) Does the table have
to contain hybridisation QC statistics? (datatype: logical)}

\item{percPres}{(Status: optional, Default:TRUE) Does the table have
to contain percentage present QC statistics? (datatype: logical)}

\item{bgPlot}{(Status: optional, Default:TRUE) Does the table have
to contain background signal intensity QC statistics? (datatype: logical)}

\item{scaleFact}{(Status: optional, Default:TRUE) Does the table have
to contain scale factor QC statistics?(datatype: logical)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}
}
\value{
A PNG file called ‘QCtable’ with the requested statistics
}
\description{
This function (from functions_imagesQC.R) computes and plots a table
of QC statistics based on the qc (simpleaffy Bioconductor package)
and yaqc (yaqcaffy Bioconductor package) functions, which generally
only work for chiptypes with perfect match and mismatch probes, but
even not for all of those. As such, when these statistics are not
provided as parameters, trys are used in this function to compute
them internally. Values for which the try fails are not computed,
but the script will continue after giving a warning.
}
\examples{
# By default, the script will call:
# computeQCtable(rawData, quality, sprep, lys, samplePrep = samplePrep,
# ratio = ratio, hybrid = hybrid, percPres = percPres, bgPlot = bgPlot, scaleFact = scaleFact)
}
