% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{readDescFile}
\alias{readDescFile}
\title{Read desc file. Return array names and groups}
\usage{
readDescFile(descfile, outprint = FALSE)
}
\arguments{
\item{descfile}{(Status=required)}

\item{outprint}{(datatype=logical, Default=FALSE)}
}
\value{
a dataframe with description matrix
}
\description{
Read desc file. Return array names and groups
}
\examples{
#example here
}
