% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{readExtFile}
\alias{readExtFile}
\title{Read norm file, depending on the file type (supports: txt, csv, xls, xlsx)}
\usage{
readExtFile(filename, extension, outprint = FALSE)
}
\arguments{
\item{filename}{(Status=required)}

\item{extension}{(Status=required)}

\item{outprint}{(datatype=logical, Default=FALSE)}
}
\value{
a dataframe with the normalized gene expression data
}
\description{
Read norm file, depending on the file type (supports: txt, csv, xls, xlsx)
}
\examples{
#example here
}
