% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{rleFun}
\alias{rleFun}
\title{Create an image with boxplots of Relative Log Expression (RLE) values}
\usage{
rleFun(
  Data,
  Data.pset = NULL,
  experimentFactor = NULL,
  plotColors = NULL,
  legendColors = NULL,
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24,
  MAXARRAY = 41
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{Data.pset}{(Status: optional, Default:NULL) An object obtained by
calling fitPLM (affyPLM), used for each but the raw plot. When not provided,
it is computed within the function.(datatype: PLMset)}

\item{experimentFactor}{(Status: required, Default:NULL) The factor of groups. (datatype: factor)}

\item{plotColors}{(Status: required, Default:NULL) Vector of colors assigned to each array. (datatype: character)}

\item{legendColors}{(Status: required, Default:NULL) Vector of colors assigned to each experimental group. (datatype: character)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}

\item{MAXARRAY}{(Status: optional, Default:41) threshold to adapt the image to the number of arrays (datatype: number)}
}
\value{
A PNG image containing boxplots of the NUSE values per array, called "RawDataRLEplot"
}
\description{
This function (from functions_imagesQC.R) creates an image with boxplots
of Relative Log Expression (RLE) values for each array. It calls the RLE
function (affyPLM Bioconductor package). An object obtained by calling
fitPLM (affyPLM) is needed. If this object is not provided, it will be
computed internally within this function.
}
\examples{
# By default, the script will call:
# rleFun(rawData, Data.pset=rawData.pset, experimentFactor, plotColors, legendColors
# where rawData.pset has been constructed by calling:
# rawData.pset <- fitPLM(rawData)
}
