% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_imagesQC.R
\name{RNAdegPlot}
\alias{RNAdegPlot}
\title{Creates an image of overall RNA degradation for all arrays}
\usage{
RNAdegPlot(
  Data,
  Data.rnadeg = NULL,
  plotColors = NULL,
  WIDTH = 1000,
  HEIGHT = 1414,
  POINTSIZE = 24,
  MAXARRAY = 41
)
}
\arguments{
\item{Data}{(Status: required) The raw data object (datatype: AffyBatch)}

\item{Data.rnadeg}{(Status: optional, Default:NULL) List as obtained
by calling the AffyRNAdeg function (affy). When not provided,
it is computed internally within this function. (datatype: list)}

\item{plotColors}{(Status: required, Default:NULL) Vector of colors assigned to each array. (datatype: character)}

\item{WIDTH}{(Status: optional, Default:1000) png image width (datatype: number)}

\item{HEIGHT}{(Status: optional, Default:1414) png image height (datatype: number)}

\item{POINTSIZE}{(Status: optional, Default:24) png image point size (datatype: number)}

\item{MAXARRAY}{(Status: optional, Default:41) threshold to adapt the image to the number of arrays (datatype: number)}
}
\value{
A PNG image of overall RNA degradation, called ‘RawDataRNAdegradationPlot’
}
\description{
This function (from functions_imagesQC.R) creates an image of
overall RNA degradation for all arrays. It calls the function
AffyRNAdeg (\href{http://www.bioconductor.org/help/bioc-views/release/bioc/html/affy.html}{affy} Bioconductor package).
}
\examples{
# RNAdegPlot(rawData,plotColors=plotColors)
}
