% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{saveComparison}
\alias{saveComparison}
\title{For each contrast, save the comparison into files}
\usage{
saveComparison(
  cont.matrix,
  fit,
  normDataTable,
  annotation = NULL,
  firstColumn = NULL
)
}
\arguments{
\item{cont.matrix}{(Status=required)}

\item{fit}{(Status=required)}

\item{normDataTable}{(Status=required)}

\item{annotation}{(default=NULL)}

\item{firstColumn}{(default=NULL)}
}
\value{
several files (for each comparison)
}
\description{
For each contrast, save the comparison into files
}
\examples{
#example here
}
