% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Stat.R
\name{toptable}
\alias{toptable}
\title{Create a toptable and save it into a file}
\usage{
toptable(contrast.fit, i, normDataTable, fileName, firstColumn, annotation)
}
\arguments{
\item{contrast.fit}{(Status=required)}

\item{i}{(Status=required) coefficient}

\item{normDataTable}{(Status=required)}

\item{fileName}{(Status=required) contrast file name}

\item{firstColumn}{(Status=required)}

\item{annotation}{(Status=required)}
}
\value{
the file name where the toptable results are saved
}
\description{
Create a toptable and save it into a file
}
\examples{
#example here
}
